+++ packages/node-core/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Update the `toBlockResponse` function signature on the worker service to take an `IBlock<B>` instead of a `B` type (#2978)
 
 ## [18.5.3] - 2025-11-25
 ### Fixed